Litt refaktorering
